Add theme/front end build tools.
First crack at adding build tools for theme and front end components.

Use the nodejs variant of pantheon's build-tools-ci container. See https://github.com/pantheon-systems/docker-build-tools-ci/pull/4 for details. This likely needs to be refined a bit.

Uses yarn for package management and gulp for building. Initial dependencies simply support that and support our custom base theme by requiring bootstrap-sass and bootswatch (in versions compatible with drupal/bootstrap, which uses the 3.x version of Bootstrap).

gulpfile.js doesn't do much to start out with, but after running the shell script that comes with the starterkit in our base theme, some basic tasks will be appended that compile the sass and copy bootstrap's js libraries. Output (sass->css and copied js) goes in the theme's build directory, which is ignored in the workflow repo, but included in the deployed version.