Add front end build steps back into circle config
Essentially reverses 3b21393b4591d09b90eaf06fbd860aa9876e3e7b. Node.js was added to Pantheon's base image in the latest release, so this I only needed to add my steps back to the build. Adding it to the functional job, which is the step that fully assembles the site to deploy and test in a pantheon environment. This may need refining, but this seems like a good place since I just need things in place when the files get synced to the Pantheon environment.