Skip visual tests when building on the same env as testing
By defaulting to dev for visual test reference site, we no longer have an untouched copy to take a reference snapshot from because the build process will have already modified it. The tests will always pass and won't necessarily be reliable. Skip to save time and output a message noting the problem.